+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/UnresolvedXMethod.java
@@ -7,7 +7,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import javax.annotation.CheckForNull;
-import javax.annotation.Nullable;
+import jakarta.annotation.Nullable;
 
 import org.apache.bcel.Const;